Move project sources under RewardFlow_API, reorganizing existing files/namespaces and Obsidian docs. Add a new Rewards feature slice (entities, DbContext, configurations, endpoints, calculators, factories and session-reward logic) plus related DTOs and interfaces. Introduce Common utilities: IEmployeeLookupService and DTOs, Optional<T> struct, EntityTracker<T>, and enMode enum. Add tests (RewardFlow_Test), solution/project files, many architecture decision records and docs, and update workspace/Obsidian settings. Remove an obsolete top-level Doc/.obsidian/app.json and the old HTTP file.

+++ src/RewardFlow.API/Employees/BulkInsertEmployees/BulkInsert.cs
@@ -18,7 +18,7 @@ public static class BulkInsert
         string Name,
         string? NationalNumber = null,
         string? AccountNumber = null,
-        float?   Salary = null);
+        decimal?   Salary = null);
     public record Request(List<emp> Employees);
     public record Response(int Success, List<int> FailsIndexes);
