+++ src/RewardFlow.API/Rewards/Common/IRewardCalculator.cs
@@ -2,5 +2,5 @@ namespace Reward_Flow_v2.Rewards.Common;
 
 public interface IRewardCalculator
 {
-    float CalculateTotal(int numOfSessions, float salary, float percentage);
+    decimal CalculateTotal(int numOfSessions, decimal salary, decimal percentage);
 }
